Add subtitle visibility toggle
Toggling subtitle visibility from the Subtitles menu and the Quicksettings panel is now available. The change is reflected by an OSD message "Subtitle Visibility: On" (or Off).

The toggle that was added in the Quicksettings is positioned at the very top in order to be visible and avoid confusion when users change subtitles and don't see any appearing on screen.

Rename the subtitle visibility toggle

The toggle is now renamed to "Hide Subtitles", the code mainly remained the same except for inverting the on for off: when subvisibility is on, the toggles are set to off, and vice versa. This change makes it more clear what this toggle does and what it means when it is On.

Revert "Rename the subtitle visibility toggle"

This reverts commit c76320503bd6d2f13fe7b1c461f8cb357512c8a6.

Rename the subtitle visibility toggle

The toggle is now "Hide Subtitles" to be more clear about what it does.