Add Editions to the playback menu
Add support for mkv editions and the ability to switch between them from
the Editions submenu under Playback. Which edition is currently selected
will be marked with a check. Which edition is default will also be
visible by appending "(Default)" to its title.
If an mkv has no editions (no chapters) or if the container does not
support editions, the Editions submenu will be empty.